+++ docs/content/misc/release_notes.rst
@@ -1,6 +1,19 @@
 Release notes
 =============
 
+0.7.x
+-----
+
+
+0.7.0
+.....
+
+- Python API:
+
+    - added n_repeats parameter to the simple TX/RX sequences available in ARRUS, the parameter allows to specify the number of times the sequence should be repeated ("batch size"),
+    - arrus.utils.imaging reconstruction operators now requires that the input data are in the format (batch size, n TX, n RX, n samples),
+    - arrus.utils.imaging.RxBeamformingImg is no longer available. Please use ReconstructLri + some form of compouding (e.g. Mean(axis=1)),
+
 0.6.x
 -----